let's go baby

+++ src/main/java/com/treno/application/dao/Dao.java
@@ -3,7 +3,7 @@ package com.treno.application.dao;
 import java.util.List;
 
 public interface Dao <T> {
-    T findById(int id);
+    T findById(long id);
     List<T> findAll();
     void save(T entity);
     void update(T entity);
